Added single page app to show reports

+++ src/Snappet.Reporting.Application/Dto/CorrectAnswersPerSubjectDto.cs
@@ -5,6 +5,6 @@ namespace Snappet.Reporting.Application.Dto
         public string Subject { get; set; }
         public int Count { get; set; }
         public int Total { get; set; }
-        public decimal Percentage => Total > 0 ? Count / Total : 0;
+        public decimal Percentage => Total > 0 ? Count / (decimal)Total : 0;
     }
 }
